- Updated Jetty to version 9.4.44.v20210927 - Updated Jackson to 2.13.0 - Updated Kafka client to 2.8.1 - Updated Vaadin to version 8.14.2 - Miscellaneous stability improvements

+++ src/main/java/org/point85/app/AppUtils.java
@@ -158,7 +158,7 @@ public abstract class AppUtils {
 			seconds = Integer.parseInt(fields[2]);
 		}
 
-		long totalSeconds = (long) hours * 3600 + (long) minutes * 60 + (long) seconds;
+		long totalSeconds = (long)hours * 3600 + minutes * 60 + seconds;
 
 		return Duration.ofSeconds(totalSeconds);
 	}
